requirements.txt
discord.py==2.3.2
openai==1.3.3
python-dotenv==1.0.0
Requests==2.31.0
